tools: add gcode-fixup
This utility measures the work area size, allows to apply offsets, scaling,
even negative one (mirroring), and will try as much as possible to optimize
for laser mode so that the head doesn't stop in raster mode each time the
intensity changes (encoded in the spindle rate). It may also scale the power
level (spindle rate), and apply a gamma correction to it (>1 = less contrasted,
<1 = more contrasted). It has no dependency beyond GNU awk which should be by
default on any Linux system. It may even be ported to standard awk after
reimplementing a few functions. It doesn't seem to use any bashism since
bash is only used to parse the command line, though if that were to be the
case this should be easy to fix.