tools: add gcode-merge-adjacent-lines.sh
This one-liner just tries to spot successive lines using the same spindle
rate and doing an X move only, then will merge them. The idea is that when
adjusting the output power level in raster mode it's possible to end up
with many lines with a similar value when some slow gradients are encoded
and their intensity becomes the same after processing. By compacting them
we reduce the amount of work for the G-CODE processor and the GRBL firmware.
For complex images at high resolutions it can even avoid some pauses.