tools/gcode-fixup: do not emit empty G0/G1 commands
Sometimes files are emitted with G0/G1 statements indicating the exact
same position as the previous one, resulting in empty G0/G1 statements.
Let's instead use a temporary string and not emit such statements if
nothing changes. This also has the benefit that we can now remove the
trailing spaces after arguments and instead prepend them if the current
line is not empty.