src: laser-preview: add a function to draw a line
The function tries to mark beam spot once per pixel and to center them
around the middle of a visited segment. Then it calls the burn() function
which tries to mark up to 4 pixels depending on its position. For now the
burn() function only marks one, as the written calculations are wrong.