src: laser-preview: implement a simplified absorption scheme
The principle is that a material has a base absorption level and an
absorption factor that depends on the amount of engraving of the
area. This allows to materials like wood which absorb energy much
more once already darkened, and others like painted aluminum which
absorb less once the paint is removed.