src: laser-preview: first attempt at implementing diffusion
The principle is to define a ratio of power diffused per pixel distance,
which is a transfer from the local point to the next one. We then compute
the linear and diagonal ratios, and since the sum of all the energy must
equal 1, we apply the inverse of the sum as a factor to all points.

Now it's pretty visible that a lot of energy is lost when diffusing to
neighbor points, we have to increase the burn power, which compares to
what has to be done to etch metals for example.