src: laser-preview: implement a simplistic gcode parser
The parser considers M3/M4/M5, G0/G1/G2/G3, S, X, and Y and draws
lines with this. It starts to work. Now it's visible that we really
need to apply some offsets to consider the material's color, which
cannot be linear.

We need to store into the area the amount of energy absorbed, which
can correspond to the amount of burn made to some wood or paint, then
we need to have a darkness() function indicating the color of a pixel.
For wood it can be something like 0 for 80% of the low energy and a
steep rise. For painted aluminum it can be 1 till 30 then a slow fade
away. This should replace the absorption and absorption factor.