tools/gcode-fixup: don't change spindle speed on M5
M5 stops the spindle but doesn't change its configured speed. For example:

  M4 S20
  M5
  M4
  G1X1

Will move to X1 with a spindle of 20. So we must not change this value there.
We might need to differentiate between "printing/not printing" and "spindle"
in the future as right now the S variable internally corresponds both to the
current spindle speed and the configured spindle speed.