src: laser-preview: use realloc() when possible in extend_img()
Most of the time when calling extend_img() we're just adding new lines
and no new column. In addition, when the lines are in the right direction
we're effectively recopying verbatim data at the same place. That's a
total waste of cycles.

Here we're detecting this situation (same width and unchanged origin)
and calling realloc() instead. Not only this allows most of the time
to reuse the same memory area, but in addition even if the area is
changed, a full memcpy() is done at once for us. This results in a 5x
speedup on large images!