src: laser-preview: fix use of negative coordinates in burn()
The function was never meant to be used with negative values due to
the way float coordinates are turned to integers. Let's instead use
floor() to convert floats to ints, and correct the distance calculations
so that the distance is always correct for each pixel's center regardless
of its sign. This also removes dx,dy which are pointless since they only
count the distance relative to (x0,y0) and do not provide a sign when it
comes to the second pixel.