src: laser-preview: make draw_vector() work with floats
The function was taking ints and was causing moire effects on some
inputs where rounding would come close to an int without touching it.
Worse, all computations were made on floats and passed to burn() which
also takes floats. Thus the degradation was really happening only when
passing the arguments. The intermediary steps are now floats as well.