src: laser-preview: make sure the total energy is stable vs offset
In burn(), when a sub-pixel offset is applied, it was possible to have
a different sum of values delivered to a group of 4 pixels due to the
way the distance was calculated. Instead we now multiply the dx*dy values
+/- 1 depending on the position, which always gives a sum of 1. The effect
is that applying sub-pixel shifts no longer blurs nor dims nor lightens
images, they're always stable.