src: laser-preview: retrieve the minimum energy density
It's now possible to specify the minimum energy density to start to
mark the material, but for now it's not used yet. The goal will be
to use it for low marking values only. It's set to 0.5J/mm^2 by
default which seems to roughly match attempts at engraving very clear
cardboard.