src: laser-preview: enforce the minimum energy density
Now we're taking this parameter into consideration, particularly for
clear areas, and it's quickly dampened for already marked ones. It
tends to show ugly images by default, but these tend to match what
we're seeing on cardboard so it makes sense. Forcing -e0 eliminates
this, of course. Some tests show that it helps find the right
parameters from png2gcode.